Simplify novpn build per review: fold inbounds/socks into boxoptions
Collapse the novpn divergence to a single build-tag surface:

- Merge inbounds_{default,novpn}.go and splittunnel_options_{default,
  novpn}.go into boxoptions_{default,novpn}.go.
- Drop the UseSocks env var and socks_only_{default,novpn}.go. The novpn
  baseInbounds now builds the SOCKS inbound directly, with a default
  address overridable by RADIANCE_SOCKS_ADDRESS.
- Fold per-platform TUN tweaks into the default baseInbounds and move
  kernelBelow/minAndroidSystemStackKernel there; OverrideAndroidVPN is
  now set in baseOpts gated on TUN-inbound presence, eliminating the
  applyPlatformTunnelOptions stub.
- Drop the bypass inbound and its routing rule from novpn via a
  build-tagged bypassRoutingRules.
