--- vpn/socks_only_default.go
@@ -1,5 +1,0 @@
-//go:build !novpn
-
-package vpn
-
-func socksOnlyEnforced() bool { return false }
